+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Change log
 
-# [RELEASED] - 2021-05-xx
+# [0.0.5] - 2021-05-12
 
 Note on the release
 
 Added
-    - added
+    - added multi-gpu support for inference
+    - added function to finetuned a model on a specific dataset on multi-gpu
 
 Changed